requirements.txt
aiogram
pandas
openpyxl
fpdf
python-dotenv
